libs/idun_agent_engine/pyproject.toml
[project]
name = "idun-agent-engine"
version = "0.6.0"
description = "Python SDK and runtime to serve AI agents with FastAPI, LangGraph, and observability."
authors = [{ name = "Geoffrey HARRAZI", email = "geoffreyharrazi@gmail.com" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = "GPL-3.0-only"
keywords = [
    "agents",
    "langgraph",
    "fastapi",
    "sdk",
    "llm",
    "observability",
]
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",



]

dependencies = [
    # Core runtime and API
    "fastapi>=0.115.0,<=0.134.0",
    "uvicorn>=0.35.0,<=0.41.0",
    "pydantic>=2.11.7,<3.0.0",
    "httpx>=0.28.1,<0.29.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "click>=8.2.0,<9.0.0",
    "platformdirs>=4.0.0,<5.0.0",
    "PyYAML>=6.0.0,<7.0.0",
    "cachetools>=5.0.0,<7.0.0",
    "httpx-sse>=0.4.0,<1.0.0",
    "botbuilder-integration-aiohttp>=4.17.0,<5.0.0",
    # Agent frameworks and orchestration
    "deepagents>=0.2.8,<1.0.0",
    "langgraph>=1.0.0,<2.0.0",
    "langgraph-prebuilt==1.0.8",
    "langgraph-checkpoint-sqlite>=3.0.0,<4.0.0",
    "langgraph-checkpoint-postgres>=3.0.0,<4.0.0",
    "google-adk>=1.19.0,<2.0.0",
    "langchain>=1.0.0,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-google-vertexai>=2.0.27,<4.0.0",
    # AG-UI and interactive interfaces
    "ag-ui-protocol==0.1.13",
    "ag-ui-langgraph==0.0.25",
    "ag-ui-adk>=0.3.4,<0.6.0",
    "copilotkit==0.1.78",
    "streamlit>=1.47.1,<2.0.0",
    "textual>=7.3.0,<8.0.0",
    # Guardrails and security
    "guardrails-ai==0.9.2",
    "openinference-instrumentation-guardrails>=0.1.0,<2.0.0",
    "PyJWT[crypto]>=2.9.0,<3.0.0",
    # Observability and telemetry
    "langsmith>=0.3.0,<1.0.0",
    "langfuse>=2.60.8,<4.0.0",
    "langfuse-haystack>=2.3.0",
    "arize-phoenix-otel>=0.2.0,<1.0.0",
    # Pin <12: 12.x currently depends on placeholder jmespath==99.99.99 and fails in Docker
    "arize-phoenix>=11.22.0,<14.0.0",
    "openinference-instrumentation-langchain>=0.1.13,<2.0.0",
    "openinference-instrumentation-vertexai>=0.1.0,<2.0.0",
    "openinference-instrumentation-google-adk>=0.1.0,<2.0.0",
    "openinference-instrumentation-mcp>=1.0.0,<2.0.0",
    "opentelemetry-exporter-gcp-trace>=1.6.0,<2.0.0",
    "opentelemetry-exporter-otlp-proto-http>=1.22.0,<2.0.0",
    "google-cloud-logging>=3.10.0,<4.0.0",
    "posthog>=7.0.0,<8.0.0",
    # Data layer and persistence
    "aiosqlite>=0.21.0,<=0.22.1",
    # Pin SQLAlchemy for Python 3.13 compatibility (required by Phoenix)
    "sqlalchemy>=2.0.36,<3.0.0",
    "psycopg-binary>=3.3.0,<4.0.0",
    # MCP and tool integrations
    "mcp>=1.0.0,<2.0.0",
    "langchain-mcp-adapters>=0.2.0,<0.3.0",
    "tavily-python>=0.7.9,<0.8.0",
    # Internal package and external delivery integrations
    "idun-agent-schema>=0.5.0,<1.0.0",
    "slack-sdk>=3.33.0,<4.0.0",
]


[project.scripts]
idun = "idun_platform_cli.main:cli"


[project.urls]
Homepage = "https://github.com/geoffreyharrazi/idun-agent-platform"
Repository = "https://github.com/geoffreyharrazi/idun-agent-platform"
Documentation = "https://github.com/geoffreyharrazi/idun-agent-platform/tree/main/libs/idun_agent_engine"
Issues = "https://github.com/geoffreyharrazi/idun-agent-platform/issues"

[dependency-groups]
dev = [
    "ruff>=0.6.5,<0.13",
    "black>=24.8.0,<27",
    "pre-commit>=3.8.0,<5",
    "mypy>=1.10.0,<2",
    "pytest>=8.3.3,<9",
    "pytest-cov>=5.0.0,<7",
    "types-PyYAML>=6.0.12.20240808,<7",
    "pytest-asyncio>=1.2.0",
    "pytest-xdist>=3.8.0",
    "posthog>=7.5.1",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/idun_agent_engine",
    "src/idun_agent_engine/py.typed",
    "src/idun_platform_cli",
]

[tool.hatch.build.targets.wheel]
include = [
  "src/idun_agent_engine",
  "src/idun_agent_engine/py.typed",
  "src/idun_platform_cli",
]

[tool.hatch.build.targets.wheel.sources]
"src/idun_agent_engine" = "idun_agent_engine"
"src/idun_platform_cli" = "idun_platform_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py312", "py313"]
include = "\\.pyi?$"
exclude = '''
(
  /build/|
  /dist/|
  /.venv/|
  /.git/
)
'''

[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = [
    "build",
    "dist",
    ".venv",
    "examples",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
]
extend-ignore = [
    "E501",  # line length handled by black
    # Keep linting actionable without behavior changes across the codebase.
    "E722",   # bare except
    "B904",   # raise from within except
    "SIM105", # prefer contextlib.suppress
    "SIM117", # combine nested with
    "SIM118", # key in dict instead of key in dict.keys()
    "UP038",  # isinstance union style
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow re-exporting in package __init__

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
disallow_untyped_defs = false
check_untyped_defs = true
pretty = true
show_error_codes = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-ra -q --strict-markers"
testpaths = ["tests/unit", "tests/integration"]
pythonpath = ["src", "tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require external services)",
    "slow: Slow tests (marked for optional skipping)",
    "requires_langfuse: Test requires Langfuse service",
    "requires_phoenix: Test requires Phoenix service",
    "requires_postgres: Test requires PostgreSQL",
    "requires_langsmith: Test requires LangSmith service",
]

[tool.coverage.run]
branch = true
source = ["src/idun_agent_engine"]
omit = [
    "*/__init__.py",
    "*/templates/*",
    "*/_version.py",
    "*/idun_agent_engine_ui/*",
    "*/idun_platform_cli/*",
    "*/mcp/*",  # CHANGED: MCP has Pydantic compatibility issues with coverage
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 0  # Start with 0, increase as coverage grows
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "pass",
]
